split <Header> into presentation & logic components
